+++ packages/sdk/src/index.ts
@@ -1,54 +1,61 @@
 import { SDK } from './sdk';
 
-export { FramePropertyNames, LayoutPropertyNames, ToolType, DownloadFormats, EnvironmentType } from './utils/enums';
+export {
+  FramePropertyNames,
+  LayoutPropertyNames,
+  ToolType,
+  DownloadFormats,
+  EnvironmentType,
+  VerticalAlign,
+  HorizontalAlign
+} from './utils/enums';
 
 export {
-    SlideDirections,
-    ShakeDirections,
-    EaseTypes,
-    TweenTypes,
-    BasicAnimationsEmphasisStyles,
+  SlideDirections,
+  ShakeDirections,
+  EaseTypes,
+  TweenTypes,
+  BasicAnimationsEmphasisStyles,
 } from './types/AnimationTypes';
 export { LayoutType, MeasurementUnit } from './types/LayoutTypes';
 export {
-    BlendMode,
-    FrameTypeEnum,
-    VerticalAlign,
-    TextDirection,
-    FlowDirection,
-    FitMode,
-    UpdateZIndexMethod,
-    ImageSourceTypeEnum,
+  BlendMode,
+  FrameTypeEnum,
+  TextDirection,
+  FlowDirection,
+  FitMode,
+  UpdateZIndexMethod,
+  ImageSourceTypeEnum,
 } from './types/FrameTypes';
 export { DocumentType } from './types/DocumentTypes';
 
 export type {
-    LayoutPropertiesType,
-    FrameProperties,
-    LayoutWithFrameProperties,
-    LayoutListItemType,
-    Layout,
+  LayoutPropertiesType,
+  FrameProperties,
+  LayoutWithFrameProperties,
+  LayoutListItemType,
+  Layout,
 } from './types/LayoutTypes';
 export type {
-    FrameLayoutType,
-    FrameType,
-    Frame,
-    TextFrame,
-    ImageFrame,
-    ShapeFrame,
-    ImageFrameSource,
-    ImageFrameVariableSource,
-    ImageFrameUrlSource,
+  FrameLayoutType,
+  FrameType,
+  Frame,
+  TextFrame,
+  ImageFrame,
+  ShapeFrame,
+  ImageFrameSource,
+  ImageFrameVariableSource,
+  ImageFrameUrlSource,
 } from './types/FrameTypes';
 export type {
-    Variable,
-    GroupVariable,
-    ImageVariable,
-    LongTextVariable,
-    ShortTextVariable,
-    ListVariable,
-    BooleanVariable,
-    ConnectorImageVariableSource,
+  Variable,
+  GroupVariable,
+  ImageVariable,
+  LongTextVariable,
+  ShortTextVariable,
+  ListVariable,
+  BooleanVariable,
+  ConnectorImageVariableSource,
 } from './types/VariableTypes';
 export { VariableType } from './types/VariableTypes';
 
@@ -56,30 +63,30 @@ export type { Color, DocumentColor, ColorUpdate } from './types/ColorStyleTypes'
 
 export type { DocumentError } from './types/DocumentTypes';
 export type {
-    FrameAnimationType,
-    FrameAnimationPropertiesType,
-    EaseTweenCombinationType,
-    AnimationPlaybackType,
-    BasicAnimationsType,
+  FrameAnimationType,
+  FrameAnimationPropertiesType,
+  EaseTweenCombinationType,
+  AnimationPlaybackType,
+  BasicAnimationsType,
 } from './types/AnimationTypes';
 export type {
-    ConfigType,
-    InitialStateType,
-    PageType,
-    EditorResponse,
-    SelectedLayoutFrame,
-    MetaData,
-    Id,
+  ConfigType,
+  InitialStateType,
+  PageType,
+  EditorResponse,
+  SelectedLayoutFrame,
+  MetaData,
+  Id,
 } from './types/CommonTypes';
 
 export type {
-    TextProperties,
-    SelectedTextStyle,
-    TextStyle,
-    AppearanceProperties,
-    TextStyleUpdateType,
-    UpdateStyleType,
-    DisplayColor,
+  TextProperties,
+  SelectedTextStyle,
+  TextStyle,
+  AppearanceProperties,
+  TextStyleUpdateType,
+  UpdateStyleType,
+  DisplayColor,
 } from './types/TextStyleTypes';
 
 export type { ParagraphStyle, ParagraphStyleUpdate } from './types/ParagraphStyleTypes';
@@ -87,20 +94,19 @@ export type { CharacterStyle, CharacterStyleUpdate } from './types/CharacterStyl
 export type { ColorUsage, ColorUsageUpdate, ColorUsageType } from './types/ColorStyleTypes';
 
 export type {
-    DocumentFontStyle,
-    DocumentFontFamily,
-    AddDocumentFontStyle,
-    AddDocumentFontFamily,
+  DocumentFontStyle,
+  DocumentFontFamily,
+  AddDocumentFontStyle,
+  AddDocumentFontFamily,
 } from './types/FontTypes';
 
 export {
-    SelectedTextStyleSections,
-    SelectedTextStyles,
-    FontWeights,
-    Alignment,
-    TextPosition,
-    Case,
-    Scripting,
+  SelectedTextStyleSections,
+  SelectedTextStyles,
+  FontWeights,
+  TextPosition,
+  Case,
+  Scripting,
 } from './types/TextStyleTypes';
 export { ColorType } from './types/ColorStyleTypes';
 export * from './types/MediaConnectorTypes';
